refactor(entity-currency-fiat): move registry constants to constants.ts
